made page id an integer
git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@565973 13f79535-47bb-0310-9956-ffa450edef68

+++ jdk-1.4/wicket/src/main/java/org/apache/wicket/Page.java
@@ -210,7 +210,7 @@ public abstract class Page extends MarkupContainer implements IRedirectListener,
 	private short autoIndex;
 
 	/** Numeric version of this page's id */
-	private short numericId;
+	private int numericId;
 
 	/** The PageMap within the session that this page is stored in */
 	private transient IPageMap pageMap;
@@ -981,7 +981,7 @@ public abstract class Page extends MarkupContainer implements IRedirectListener,
 	 */
 	public final void setNumericId(final int id)
 	{
-		numericId = (short)id;
+		numericId = id;
 	}
 
 	/**
